feat: implement `EXCSEntity` base class and refactor `EXCSTracksPowerSwitch` for improved structure
